% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumida.R
\name{cumida.get}
\alias{cumida.get}
\title{Download from CuMiDa}
\usage{
cumida.get(gse, tissue, platform = TRUE, ..., verbose = TRUE)
}
\arguments{
\item{gse}{(\code{character} or \code{integer}) : The GSE accession ID}

\item{tissue}{(\code{character}) : The target tissue name (case-insensitive)}

\item{platform}{(\code{logical}) : if \code{TRUE}, the platform data describing each
probe is also downloaded and included in the dataset}

\item{...}{optional arguments (for compatibility only)}

\item{verbose}{(\code{logical}) : whether to print each step (\code{TRUE}) or run
silently (\code{FALSE})}
}
\value{
An \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}} object
}
\description{
Functions to download and process microarray data
from the CuMiDa repository.
}
\details{
\code{cumida.get} downloads a microarray dataset
from the CuMiDa repository and processes into an \code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}}.
}
