% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancada-package.R
\docType{package}
\name{pancada-package}
\alias{pancada}
\alias{pancada-package}
\title{pancada: PanCaDa}
\description{
R Package to querying, fetching, and analysis of datasets from Pan Cancer Database (PanCaDa).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nunesijg/pancada}
}

}
\author{
\strong{Maintainer}: Itamar José Guimarães Nunes \email{nunesijg@gmail.com} (\href{https://orcid.org/0000-0002-6246-4658}{ORCID})

Other contributors:
\itemize{
  \item Bruno César Feltes \email{bcfeltes@gmail.com} (\href{https://orcid.org/0000-0002-2825-8295}{ORCID}) [contributor]
  \item Joice de Faria Poloni \email{joicefpoloni@gmail.com} (\href{https://orcid.org/0000-0002-3896-6404}{ORCID}) [contributor]
  \item Marcio Dorn \email{mdorn@inf.ufrgs.br} (\href{https://orcid.org/0000-0001-8534-3480}{ORCID}) [contributor]
}

}
\keyword{internal}
